Refactor only - Move Compaction related logic in HoodieWriteClient to separate handler class but retain all public APIs
